Remove trait for large value persistence
